backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
passlib==1.7.4
bcrypt==4.0.1
python-dotenv
pymongo
aiohttp
requests
groq
reportlab
PyMuPDF
PyPDF2
